requirement.txt
asgiref==3.7.2
Babel==2.13.1
Django==4.2.6
django-autocomplete-light==3.9.7
django-js-asset==2.1.0
django-mptt==0.15.0
django-phonenumber-field==7.2.0
django-querysetsequence==0.17
django-simpleui==2023.8.28
phonenumbers==8.13.23
Pillow==10.1.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
